Moving parallel functionality into aat.py.
