+++ magpy/tests/test_024_AAT_MP2_parallel.py
@@ -28,7 +28,7 @@ def test_AAT_MP2_H2DIMER():
     b_disp = 0.0001
     e_conv = 1e-12
     r_conv = 1e-12
-    I_00, I_0D, I_D0, I_DD = AAT.compute('MP2', r_disp, b_disp, e_conv=e_conv, r_conv=r_conv, normalization='intermediate', parallel=True)
+    I_00, I_0D, I_D0, I_DD = AAT.compute('MP2', r_disp, b_disp, e_conv=e_conv, r_conv=r_conv, normalization='intermediate', parallel=True, num_procs=16)
     print("\nElectronic Contribution to Atomic Axial Tensor (a.u.):")
     print("Hartree-Fock component:")
     print(I_00)
@@ -126,7 +126,7 @@ def test_AAT_MP2_H2DIMER_NORM():
     b_disp = 0.0001
     e_conv = 1e-12
     r_conv = 1e-12
-    I_00, I_0D, I_D0, I_DD = AAT.compute('MP2', r_disp, b_disp, e_conv=e_conv, r_conv=r_conv, normalization='full', parallel=True)
+    I_00, I_0D, I_D0, I_DD = AAT.compute('MP2', r_disp, b_disp, e_conv=e_conv, r_conv=r_conv, normalization='full', parallel=True, num_procs=16)
 
     print("\nElectronic Contribution to Atomic Axial Tensor (a.u.):")
     print("Hartree-Fock component:")
@@ -225,7 +225,7 @@ def test_AAT_MP2_H2O():
     b_disp = 0.0001
     e_conv = 1e-12
     r_conv = 1e-12
-    I_00, I_0D, I_D0, I_DD = AAT.compute('MP2', r_disp, b_disp, e_conv=e_conv, r_conv=r_conv, normalization='intermediate', parallel=True)
+    I_00, I_0D, I_D0, I_DD = AAT.compute('MP2', r_disp, b_disp, e_conv=e_conv, r_conv=r_conv, normalization='intermediate', parallel=True, num_procs=16)
     print("\nElectronic Contribution to Atomic Axial Tensor (a.u.):")
     print("Hartree-Fock component:")
     print(I_00)
@@ -311,7 +311,7 @@ def test_AAT_MP2_H2O_NORM():
     b_disp = 0.0001
     e_conv = 1e-12
     r_conv = 1e-12
-    I_00, I_0D, I_D0, I_DD = AAT.compute('MP2', r_disp, b_disp, e_conv=e_conv, r_conv=r_conv, normalization='full', parallel=True)
+    I_00, I_0D, I_D0, I_DD = AAT.compute('MP2', r_disp, b_disp, e_conv=e_conv, r_conv=r_conv, normalization='full', parallel=True, num_procs=16)
 
     print("\nElectronic Contribution to Atomic Axial Tensor (a.u.):")
     print("Hartree-Fock component:")
